+++ R/get_predicted.R
@@ -85,10 +85,11 @@
 #' @param ci_method The method for computing p values and confidence intervals.
 #'   Possible values depend on model type.
 #'   + `NULL` uses the default method, which varies based on the model type.
-#'   + Most frequentist models: `"normal"` (default).
+#'   + Most frequentist models: `"wald"` (default) or `"normal"`.
 #'   + Bayesian models:  `"quantile"`  (default), `"hdi"`, `"eti"`, and `"spi"`.
-#'   + Mixed effects **lme4** models: `"normal"` (default), `"satterthwaite"`, and
-#'     `"kenward"`.
+#'   + Mixed effects **lme4** models: `"wald"` (default), `"normal"`,
+#'   `"satterthwaite"`, and `"kenward-roger"`.
+#'   See [`get_df()`] for details.
 #' @param dispersion_method Bootstrap dispersion and Bayesian posterior summary:
 #'   `"sd"` or `"mad"`.
 #' @param ... Other argument to be passed, for instance to `get_predicted_ci()`.