+++ tests/testthat/test-glmmTMB.R
@@ -86,6 +86,29 @@ if (requiet("testthat") &&
     expect_equal(get_df(m2, type = "model"), 4)
   })
 
+  test_that("get_df", {
+    expect_equal(
+      get_df(m2, type = "residual"),
+      df.residual(m2),
+      ignore_attr = TRUE
+    )
+    expect_equal(
+      get_df(m2, type = "normal"),
+      Inf,
+      ignore_attr = TRUE
+    )
+    expect_equal(
+      get_df(m2, type = "wald"),
+      Inf,
+      ignore_attr = TRUE
+    )
+    expect_equal(
+      get_df(m2, type = "ml1"),
+      c(`(Intercept)` = 247, child = 247, camper1 = 247),
+      ignore_attr = TRUE
+    )
+  })
+
   test_that("model_info", {
     expect_true(model_info(m1)$is_zero_inflated)
     expect_false(model_info(m2)$is_zero_inflated)