+++ tests/testthat/test-lm.R
@@ -62,6 +62,24 @@ if (requiet("testthat") &&
     expect_equal(get_df(m2, type = "model"), attr(logLik(m2), "df"), ignore_attr = TRUE)
   })
 
+  test_that("get_df", {
+    expect_equal(
+      get_df(m1, type = "residual"),
+      df.residual(m1),
+      ignore_attr = TRUE
+    )
+    expect_equal(
+      get_df(m1, type = "normal"),
+      Inf,
+      ignore_attr = TRUE
+    )
+    expect_equal(
+      get_df(m1, type = "wald"),
+      df.residual(m1),
+      ignore_attr = TRUE
+    )
+  })
+
   test_that("get_data", {
     expect_equal(nrow(get_data(m1)), 150)
     expect_equal(