+++ tests/testthat/test-polr.R
@@ -35,6 +35,24 @@ if (requiet("testthat") &&
     )
   })
 
+  test_that("get_df", {
+    expect_equal(
+      get_df(m1, type = "residual"),
+      df.residual(m1),
+      ignore_attr = TRUE
+    )
+    expect_equal(
+      get_df(m1, type = "normal"),
+      Inf,
+      ignore_attr = TRUE
+    )
+    expect_equal(
+      get_df(m1, type = "wald"),
+      df.residual(m1), # model has t-statistic
+      ignore_attr = TRUE
+    )
+  })
+
   test_that("find_formula", {
     expect_length(find_formula(m1), 1)
     expect_equal(
@@ -135,7 +153,7 @@ if (requiet("testthat") &&
     expect_true(all(c("Row", "Response", "Predicted") %in% colnames(p1)))
     expect_true(all(c("Row", "Response", "Predicted") %in% colnames(p3)))
 
-    d <- get_datagrid(m1, at = "Type")
+    d <- get_datagrid(m1, at = "Type", verbose = FALSE)
     expect_warning(get_predicted(m1, predict = "expectation", data = d))
     p1 <- suppressWarnings(get_predicted(m1, predict = "expectation", data = d, verbose = FALSE))
     expect_equal(colnames(p1), c("Row", "Type", "Response", "Predicted"))